refactor: extract StartCardAnimations from StartLayoutAnimator
