+++ src/main/java/edu/ntnu/idatt2003/millions/view/start/StartLayoutAnimator.java
@@ -2,98 +2,27 @@ package edu.ntnu.idatt2003.millions.view.start;
 
 import edu.ntnu.idatt2003.millions.view.StartView;
 import edu.ntnu.idatt2003.millions.view.component.AppTabPane;
-import javafx.animation.Interpolator;
-import javafx.animation.KeyFrame;
-import javafx.animation.KeyValue;
-import javafx.animation.PauseTransition;
-import javafx.animation.Timeline;
-import javafx.application.Platform;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.DoubleBinding;
 import javafx.scene.Node;
-import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
-import javafx.util.Duration;
 
 /**
- * Coordinates start-screen layout binding and reveal animations.
+ * Coordinates start-screen layout binding and tab-height listeners.
  *
- * <p>This component keeps JavaFX animation and responsive height binding logic
- * out of {@link StartView} and {@link NewGameTab}. It owns only view concerns:
- * the automatic downward upload reveal, and the responsive card frame that
- * lets the active tab resize cleanly.</p>
- *
- * <p>Construction of the collapsed upload section is delegated to
- * {@link UploadSectionFactory}.</p>
+ * <p>Wires responsive height bindings, upload-reveal tracking, and feedback
+ * height listeners for the start card. All {@link Timeline}-based animations
+ * are delegated to {@link StartCardAnimations}, and construction of the
+ * collapsed upload section is delegated to {@link UploadSectionFactory}.</p>
  */
 public final class StartLayoutAnimator {
 
     private static final double START_TITLE_ESTIMATED_HEIGHT = 72;
     private static final double START_GROUP_HEIGHT_RATIO = 0.72;
     private static final double START_TAB_HEADER_HEIGHT = 60;
-    private static final double UPLOAD_SECTION_MIN_HEIGHT = 200;
-    private static final double UPLOAD_SECTION_BUFFER = 8;
-    private static final double CONTENT_WIDTH_EXTRA = 48;
-    private static final Duration UPLOAD_REVEAL_DELAY = Duration.millis(380);
-    private static final Duration UPLOAD_REVEAL_DURATION = Duration.millis(820);
 
     private StartLayoutAnimator() {
-    }
-
-    /**
-     * Plays the automatic upload reveal for {@link NewGameTab}.
-     *
-     * <p>The upload section expands downward while the owning tab content
-     * animates its own preferred height, allowing the surrounding {@link AppTabPane}
-     * to resize the white start card during the reveal.</p>
-     *
-     * @param tabContent    the tab content whose height should follow the reveal
-     * @param uploadSection the collapsed upload section to reveal
-     * @param cardWidth     the visual card content width
-     */
-    public static void playUploadIntro(VBox tabContent, VBox uploadSection, double cardWidth) {
-        Platform.runLater(() -> {
-            tabContent.applyCss();
-            tabContent.layout();
-            double collapsedHeight = tabContent.prefHeight(cardWidth + CONTENT_WIDTH_EXTRA);
-            double uploadExpandedHeight = computeUploadSectionExpandedHeight(uploadSection, cardWidth);
-            double expandedHeight = collapsedHeight + uploadExpandedHeight;
-
-            tabContent.setMinHeight(collapsedHeight);
-            tabContent.setPrefHeight(collapsedHeight);
-            tabContent.setMaxHeight(expandedHeight);
-
-            PauseTransition delay = new PauseTransition(UPLOAD_REVEAL_DELAY);
-            delay.setOnFinished(event -> playUploadTimeline(
-                    tabContent, uploadSection, collapsedHeight, expandedHeight, uploadExpandedHeight));
-            delay.play();
-        });
-    }
-
-    /**
-     * Computes the natural expanded height of {@code uploadSection} by summing
-     * the preferred heights of its {@link Region} children and the spacing
-     * between them, plus a small safety buffer to absorb font and rendering
-     * rounding so the section's clip never cuts visible content.
-     *
-     * @param uploadSection the upload section whose children determine the height
-     * @param width         the width used to compute each child's preferred height
-     * @return the target expanded height in pixels
-     */
-    private static double computeUploadSectionExpandedHeight(VBox uploadSection, double width) {
-        double sum = 0;
-        int count = 0;
-        for (Node child : uploadSection.getChildrenUnmodifiable()) {
-            if (child instanceof Region region) {
-                sum += region.prefHeight(width);
-                count++;
-            }
-        }
-        if (count > 1) {
-            sum += uploadSection.getSpacing() * (count - 1);
-        }
-        return Math.max(UPLOAD_SECTION_MIN_HEIGHT, sum + UPLOAD_SECTION_BUFFER);
     }
 
     /**
@@ -127,7 +56,7 @@ public final class StartLayoutAnimator {
                                            double cardWidth,
                                            double groupSpacing) {
         // Set initial tab pane height before upload animation starts
-        double initialHeight = newGameTab.prefHeight(cardWidth + CONTENT_WIDTH_EXTRA)
+        double initialHeight = newGameTab.prefHeight(cardWidth + StartCardAnimations.CONTENT_WIDTH_EXTRA)
                 + START_TAB_HEADER_HEIGHT;
         tabPane.setMinHeight(initialHeight);
         tabPane.setPrefHeight(initialHeight);
@@ -150,8 +79,9 @@ public final class StartLayoutAnimator {
             if (newTab == null) return;
             double target = newTab.getContent() == newGameTab
                     ? newGameTab.getPrefHeight() + START_TAB_HEADER_HEIGHT
-                    : loadGameTab.prefHeight(cardWidth + CONTENT_WIDTH_EXTRA) + START_TAB_HEADER_HEIGHT;
-            playTabHeightTransition(tabPane, target);
+                    : loadGameTab.prefHeight(cardWidth + StartCardAnimations.CONTENT_WIDTH_EXTRA)
+                            + START_TAB_HEADER_HEIGHT;
+            StartCardAnimations.playTabHeightTransition(tabPane, target);
         });
 
         // Drive tab-content height explicitly from each tab's feedbackVisibleProperty
@@ -165,63 +95,6 @@ public final class StartLayoutAnimator {
                 center, newGameTab, cardWidth, groupSpacing));
         reservedStartGroup.minHeightProperty().bind(reservedStartGroup.prefHeightProperty());
         reservedStartGroup.maxHeightProperty().bind(reservedStartGroup.prefHeightProperty());
-    }
-
-    private static void playUploadTimeline(VBox tabContent,
-                                           VBox uploadSection,
-                                           double collapsedHeight,
-                                           double expandedHeight,
-                                           double uploadExpandedHeight) {
-        Timeline reveal = new Timeline(
-                new KeyFrame(Duration.ZERO,
-                        new KeyValue(uploadSection.prefHeightProperty(), 0),
-                        new KeyValue(uploadSection.maxHeightProperty(), 0),
-                        new KeyValue(uploadSection.opacityProperty(), 0),
-                        new KeyValue(tabContent.minHeightProperty(), collapsedHeight),
-                        new KeyValue(tabContent.prefHeightProperty(), collapsedHeight)),
-                new KeyFrame(UPLOAD_REVEAL_DURATION,
-                        new KeyValue(uploadSection.prefHeightProperty(),
-                                uploadExpandedHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(uploadSection.maxHeightProperty(),
-                                uploadExpandedHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(uploadSection.opacityProperty(), 1, Interpolator.EASE_OUT),
-                        new KeyValue(tabContent.minHeightProperty(), expandedHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(tabContent.prefHeightProperty(), expandedHeight, Interpolator.EASE_BOTH))
-        );
-        reveal.setOnFinished(finished -> {
-            uploadSection.setPrefHeight(uploadExpandedHeight);
-            uploadSection.setMaxHeight(uploadExpandedHeight);
-            uploadSection.setOpacity(1);
-            tabContent.setMinHeight(expandedHeight);
-            tabContent.setPrefHeight(expandedHeight);
-            // Leave maxHeight unconstrained so the feedback label can push height upward.
-            tabContent.setMaxHeight(Double.MAX_VALUE);
-        });
-        reveal.play();
-    }
-
-    /**
-     * Animates the tab pane height from its current value to {@code targetHeight}.
-     *
-     * @param tabPane      the tab pane to animate
-     * @param targetHeight the target height in pixels
-     * @return the running {@link Timeline}, so callers can stop it if a follow-up
-     *         animation needs to override it before completion
-     */
-    private static Timeline playTabHeightTransition(AppTabPane tabPane, double targetHeight) {
-        double fromHeight = tabPane.getPrefHeight();
-        Timeline timeline = new Timeline(
-                new KeyFrame(Duration.ZERO,
-                        new KeyValue(tabPane.prefHeightProperty(), fromHeight),
-                        new KeyValue(tabPane.minHeightProperty(), fromHeight),
-                        new KeyValue(tabPane.maxHeightProperty(), fromHeight)),
-                new KeyFrame(Duration.millis(350),
-                        new KeyValue(tabPane.prefHeightProperty(), targetHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(tabPane.minHeightProperty(), targetHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(tabPane.maxHeightProperty(), targetHeight, Interpolator.EASE_BOTH))
-        );
-        timeline.play();
-        return timeline;
     }
 
     /**
@@ -252,9 +125,9 @@ public final class StartLayoutAnimator {
                     ? baseHeight[0] + tab.computeFeedbackReservedHeight()
                     : baseHeight[0];
             if (isNewGame) {
-                active[0] = playTabContentHeightTransition(tab, target);
+                active[0] = StartCardAnimations.playTabContentHeightTransition(tab, target);
             } else if (isLoadTabActive(tabPane, tab)) {
-                active[0] = playTabHeightTransition(tabPane, target);
+                active[0] = StartCardAnimations.playTabHeightTransition(tabPane, target);
             }
         });
     }
@@ -271,33 +144,6 @@ public final class StartLayoutAnimator {
         Node selected = tabPane.getSelectionModel().getSelectedItem() == null
                 ? null : tabPane.getSelectionModel().getSelectedItem().getContent();
         return selected == loadTab;
-    }
-
-    /**
-     * Animates the {@code min}, {@code pref} and {@code max} height of a tab
-     * content {@link FileDropTab} from its current preferred height to
-     * {@code targetHeight}. Sister method of {@link #playTabHeightTransition}
-     * but operates on the content {@code VBox} instead of the tab pane.
-     *
-     * @param content      the tab content to animate
-     * @param targetHeight the target height in pixels
-     * @return the running {@link Timeline}, so callers can stop it if a follow-up
-     *         animation needs to override it before completion
-     */
-    private static Timeline playTabContentHeightTransition(FileDropTab content, double targetHeight) {
-        double fromHeight = content.getPrefHeight();
-        Timeline timeline = new Timeline(
-                new KeyFrame(Duration.ZERO,
-                        new KeyValue(content.prefHeightProperty(), fromHeight),
-                        new KeyValue(content.minHeightProperty(), fromHeight),
-                        new KeyValue(content.maxHeightProperty(), fromHeight)),
-                new KeyFrame(Duration.millis(250),
-                        new KeyValue(content.prefHeightProperty(), targetHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(content.minHeightProperty(), targetHeight, Interpolator.EASE_BOTH),
-                        new KeyValue(content.maxHeightProperty(), targetHeight, Interpolator.EASE_BOTH))
-        );
-        timeline.play();
-        return timeline;
     }
 
     /**
@@ -321,7 +167,7 @@ public final class StartLayoutAnimator {
                     // Always use newGameTab height to keep title and card position stable
                     double contentHeight = START_TITLE_ESTIMATED_HEIGHT
                             + groupSpacing
-                            + newGameTab.prefHeight(cardWidth + CONTENT_WIDTH_EXTRA)
+                            + newGameTab.prefHeight(cardWidth + StartCardAnimations.CONTENT_WIDTH_EXTRA)
                             + START_TAB_HEADER_HEIGHT;
                     return Math.min(availableHeight, Math.max(proportionalHeight, contentHeight));
                 },
